Перемещение компоненты создания заявки в главную страницу
